api/tests/requirements-test.txt
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
python-dotenv>=1.0
python-jose[cryptography]>=3.3
